Some fixes for SDL device
- touch input id is always unsigned 64bit
- no need to check relative mouse mode and generate keymap each time
(in case when null device is used)
- set simulated mouse/touch events after SDL is initialized (at least on linux
it stopped working when it was set before sdl init)

+++ include/IEventReceiver.h
@@ -439,7 +439,7 @@ struct SEvent
 	struct STouchInput
 	{
 		// Touch ID.
-		size_t ID;
+		u64 ID;
 
 		// X position of simple touch.
 		f64 X;
